fix: linter fix

+++ src/ChainLinkOracleWrapper.sol
@@ -4,20 +4,15 @@ pragma solidity ^0.8.20;
 import { IChainlinkAggregatorV3Interface } from "src/interfaces/deps/IChainlinkAggregatorV3Interface.sol";
 
 contract ChainLinkOracleWrapper {
-    constructor() {}
+    constructor() { }
 
     /**
      * Returns the latest price
      */
-    function getLatestPrice(address priceFeedAddress) public view returns (uint80, int, uint, uint, uint80) {
+    function getLatestPrice(address priceFeedAddress) public view returns (uint80, int256, uint256, uint256, uint80) {
         IChainlinkAggregatorV3Interface priceFeed = IChainlinkAggregatorV3Interface(priceFeedAddress);
-        (
-            uint80 roundID,
-            int price,
-            uint startedAt,
-            uint timeStamp,
-            uint80 answeredInRound
-        ) = priceFeed.latestRoundData();
+        (uint80 roundID, int256 price, uint256 startedAt, uint256 timeStamp, uint80 answeredInRound) =
+            priceFeed.latestRoundData();
         return (roundID, price, startedAt, timeStamp, answeredInRound);
     }
 }
